Schema refactorings and tidying update
The Sender::Payload and Receiver::Payload models are now the leading objects
when interacting with payloads.

When the app is configured to be the sender, Sender::Payload.receive! can be
called when the payload was received from the receiver (lol)
When the app is configured to be the receiver, Receiver::Payload.send! can be
called to send the current status of the storage to the sender.

Furthermore, the sync is now connected to the payload one-to-one in the
database. Also moved the has_one_attached to the payload model instead of the
sync model.