Use normal bigint ID for payload primary keys
The issue I was seeing was pretty funny because it was also flaky: When
activestorage tried to attach the files, it was luck if it would actually save
it because the primary key was a UID (SecureRandom.hex(6)) and when that was
generated with an integer in the beginning, it converted it to a valid entry in
the active storage table. So for simplicity's sake let's use normal bigint IDs everywhere.