Restructure payload
There was a conceptual issue how we diffed the files. Of course the mtime
can (and will be) newer on the receiver because the write will happen later on
the receiver's side which results in a newer timestamp than on the sender's file system.

With the new payload, we now use a hash and use the key as a timestamp instead
of a string, which is more efficient anyway. This way the mtime of the files in
the payload can be directly accessed.

Now the comparison is done the following way: If the payload's mtime is older,
then it needs to be re-sent. Before it was a 1-1 comparison, which expected the
files to have the exact same mtime on the two different systems. Which is
obviously never true.